♻️ refactor: enums.py 코드 통합
users 디렉토리 enums.py 삭제 후 users/utils 디렉토리의 enums.py로 코드 통합, 코드 통합에 따른 import 구문 수정
